Fix KeyError when using Run or Run Configuration action
Fix KeyError when Run or "Run Configuration per file" actions are
triggered after the following steps:

  - Close all files so that Spyder crates a new untitled file

  - Make sure _not_ to click inside the editor (the "input field", where
    the mouse cursor changes to the I-beam cursor) or an menu entry in
    the main menu. However, you may click other widgets, including the
    editors tab- and title-bar.

  - Now trigger the Run or "Run Configuration per file" action using the
    corresponding button in the toolbar or the corresponding shortcuts.

The Run Configuration dialog relies on "focus changed" signals emitted
by the editor to know which file is currently focused and thereby which
configuration to show/to apply on run. When the last editor is closed,
such an event is lost.

Fix the issue by making sure that the "focus changed" signal is emitted
when a new (and only) file is created.